Changed clock field type to BigInteger (closes #241)

+++ migasfree/server/models/hw_node.py
@@ -160,7 +160,7 @@ class HwNode(models.Model, MigasLink):
         null=True
     )
 
-    clock = models.IntegerField(
+    clock = models.BigIntegerField(
         verbose_name=_("clock"),
         null=True
     )
